Move fixed tests to done
