+++ src/test/java/com/powsybl/openloadflow/ac/knitro/done/NonImpedantBranchDisablingTest.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.openloadflow.ac.knitro.tofix;
+package com.powsybl.openloadflow.ac.knitro.done;
 
 import com.powsybl.commons.report.ReportNode;
 import com.powsybl.iidm.network.Network;
